+++ src/openrct2/libopenrct2.vcxproj
@@ -276,6 +276,7 @@
     <ClInclude Include="paint\tile_element\Paint.TileElement.h" />
     <ClInclude Include="paint\VirtualFloor.h" />
     <ClInclude Include="ParkImporter.h" />
+    <ClInclude Include="peep\GuestPathfinding.h" />
     <ClInclude Include="peep\Peep.h" />
     <ClInclude Include="peep\Staff.h" />
     <ClInclude Include="PlatformEnvironment.h" />
